% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{add_catmaply_slider}
\alias{add_catmaply_slider}
\title{Add catmaply slider traces}
\usage{
add_catmaply_slider(
  fig,
  df,
  text_color = "#444",
  slider_prefix = "",
  slider_steps,
  hover_hide = NA,
  color_palette,
  categorical_colorbar,
  category_items,
  legend_items
)
}
\arguments{
\item{fig}{plotly object}

\item{df}{data.frame or tibble holding the data.}

\item{text_color}{font color to be used for text; (default: "#444").}

\item{slider_steps}{list holding the configuration of the steps to be created. There are two alternatives: \code{auto} and
\code{custom}; whereas the \code{auto} mode creates the steps automatically and \code{custom} takes custom instructions on how to create the steps.
For mode \code{auto}, a \code{list} with the following elements has to be submitted (values of the list element are just examples): \cr
list( \cr
  slider_start=1, \cr
  slider_range=15, \cr
  slider_shift=5, \cr
) \cr
This will create the steps automatically for you, essentially starting at position \code{slider_start},
shifting the window of size \code{slider_range} along the x axis with a stepsize of \code{slider_shift}. The stepnames
are automatically selected with the x value of the left side of the slider_range (so for 1 it would take the first value of the x axis as name of the step). \cr
With custom, on the other hand, you can define the step configuration without any restrictions. The custom
configuration needs to be defined in a \code{list} with the following elements. \cr
list( \cr
  list(name="Step_One", range=c(1, 50)), \cr
  list(name="Step_Two", range=c(5, 55)), \cr
  ... \cr
). \cr
(default: \cr
list( \cr
  slider_start=1, \cr
  slider_range=15, \cr
  slider_shift=5, \cr
)).}

\item{hover_hide}{boolean indicating if the hover label should be hidden or not; (default: FALSE).}

\item{color_palette}{a color palette vector.}

\item{categorical_colorbar}{if the resulting heatmap holds categorical field values or continuous values that belong to a category; (default: FALSE).}

\item{category_items}{disctinct/unique items of ordered category items}

\item{legend_items}{distinct/unique items of ordered legend items}

\item{categorical_col}{if categorical_colorbar is TRUE, then this column is used to create categories; (default: FALSE).}
}
\value{
plot_ly object
}
\description{
Function to produce catmaply traces.
}
\keyword{internal}
