% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.R
\name{bt}
\alias{bt}
\title{Bt with discrete colorbar}
\usage{
bt(df, annotated = F)
}
\arguments{
\item{df}{input data}

\item{annotated}{True if you want to add annotations to the plot (Default: False)}
}
\value{
plotly object
}
\description{
Bt with discrete colorbar
}
