% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaply.R
\docType{package}
\name{catmaply-package}
\alias{catmaply}
\alias{catmaply-package}
\title{catmaply: Heatmap for categorical data using plotly}
\description{
To learn more about catmaply, start with the vignettes:
`browseVignettes(package = "catmaply")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yvesmauron/catmaply}
  \item Report bugs at \url{https://github.com/yvesmauron/catmaply/issues}
}

}
\author{
\strong{Maintainer}: Yves Mauron \email{yves.mauron@trivadis.com} [copyright holder]

Authors:
\itemize{
  \item Aron Weller
}

}
\keyword{internal}
