% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.R
\name{catmaply}
\alias{catmaply}
\title{catmaply with discrete colorbar}
\usage{
catmaply(df, annotated = F)
}
\arguments{
\item{df}{input data}

\item{annotated}{True if you want to add annotations to the plot (Default: False)}
}
\value{
plotly object
}
\description{
catmaply with discrete colorbar
}
