% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{catmaply_annotations}
\alias{catmaply_annotations}
\title{Get catmaply annotation list}
\usage{
catmaply_annotations(
  df,
  annotated,
  text_color = "#444",
  text_size = 12,
  text_font_family = c("Open Sans", "verdana", "arial", "sans-serif")
)
}
\arguments{
\item{df}{data.frame or tibble holding the data.}

\item{annotated}{boolean indicating if annotations should be displayed.}

\item{text_color}{font color to be used for text; (default: "#444").}

\item{text_size}{font size to be used for text/annotation. Needs to be a number greater than or equal to 1; (default: 12).}
}
\value{
list
}
\description{
Function to produce catmaply traces.
}
\keyword{internal}
