% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.R
\name{catmaply_demo}
\alias{catmaply_demo}
\title{catmaply Demo Data}
\usage{
catmaply_demo(input_data)
}
\arguments{
\item{input_data}{input_data from generate_demo_data}
}
\value{
plotly object
}
\description{
to be removed
}
