% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.R
\name{catmaply_trace_time}
\alias{catmaply_trace_time}
\title{catmaply with traces and time}
\usage{
catmaply_trace_time(df)
}
\arguments{
\item{df}{input data}
}
\value{
plotly object
}
\description{
To be merged with catmaply_traces
}
