% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrete_coloring_range}
\alias{discrete_coloring_range}
\title{Generates the parameters necessary for discrete coloring and colorbar}
\usage{
discrete_coloring_range(df, color_palette, categorical_color_range)
}
\arguments{
\item{df}{catmaply tibble}

\item{categorical_color_range}{whether to use categorical color range or not.}

\item{col_palette}{the color palette}
}
\value{
list(colorscale, tickvals, ticktext)
}
\description{
Generates the parameters necessary for discrete coloring and colorbar
}
\keyword{internal}
