% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate Test Data}
\usage{
generate_test_data(nr_stops = 35, nr_drives = 100)
}
\arguments{
\item{nr_stops}{number of stops}

\item{nr_drives}{number of drives}
}
\value{
tibble
}
\description{
Generate Test Data
}
