% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vbz}
\alias{vbz}
\title{Sample files provided by VBZ}
\format{
list in the following structure
\describe{
  \item{file_name}{original file name}
  \item{data}{data.frame holding the data}
  ...
}
}
\source{
vbz
}
\usage{
vbz
}
\description{
Sample data
}
\keyword{datasets}
