% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaply.R
\docType{package}
\name{catmaply-package}
\alias{catmaply-package}
\alias{_PACKAGE}
\title{catmaply: Heatmap for Categorical Data using Plotly}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides utility and plotting functions to efficiently plot categorical data on a heatmap using plotly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yvesmauron/catmaply}
  \item Report bugs at \url{https://github.com/yvesmauron/catmaply/issues}
}

}
\author{
\strong{Maintainer}: Yves Mauron \email{ypmauron@gmail.com}

Authors:
\itemize{
  \item Aron Weller
}

}
\keyword{internal}
