% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrete_coloring}
\alias{discrete_coloring}
\title{Generates the parameters necessary for discrete coloring and colorbar}
\usage{
discrete_coloring(df, color_palette, categorical_color_range, legend_items)
}
\arguments{
\item{df}{catmaply tibble}

\item{color_palette}{the color palette}

\item{categorical_color_range}{whether to use categorical color range or not.}

\item{legend_items}{distinct/unique items of ordered legend items}
}
\value{
list(colorscale, tickvals, ticktext)
}
\description{
Generates the parameters necessary for discrete coloring and colorbar
}
\keyword{internal}
