% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discrete_coloring}
\alias{discrete_coloring}
\title{Generates the parameters necessary for dicrete coloring and colorbar}
\usage{
discrete_coloring(categories, col_palette)
}
\arguments{
\item{categories}{categories, for coloring should done}

\item{col_palette}{the color palette}
}
\value{
list(colorscale, tickvals, ticktext)
}
\description{
Generates the parameters necessary for dicrete coloring and colorbar
}
\keyword{internal}
